requirements.txt
pydantic==2.1.1
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
python-telegram-bot==20.4
vk-api==11.9.9
SQLAlchemy==2.0.19
